chore: fixed column issues and add column configuration to tsx
